Pull water effect code out of renderer
