+++ src/engine/graphical_effects.hpp
@@ -1,4 +1,4 @@
-/* Copyright (C) 2021, Nikolai Wuttke. All rights reserved.
+/* Copyright (C) 2022, Nikolai Wuttke. All rights reserved.
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -16,14 +16,42 @@
 
 #pragma once
 
+#include "renderer/custom_quad_batch.hpp"
 #include "renderer/shader.hpp"
+#include "renderer/texture.hpp"
 
 
 namespace rigel::renderer
 {
+class Renderer;
+}
 
-extern const ShaderSpec TEXTURED_QUAD_SHADER;
-extern const ShaderSpec SIMPLE_TEXTURED_QUAD_SHADER;
-extern const ShaderSpec SOLID_COLOR_SHADER;
+namespace rigel::engine
+{
 
-} // namespace rigel::renderer
+struct WaterEffectArea
+{
+  base::Rect<int> mArea;
+  bool mIsAnimated;
+};
+
+
+class WaterEffectRenderer
+{
+public:
+  WaterEffectRenderer(renderer::Renderer* pRenderer);
+
+  void draw(
+    const renderer::RenderTargetTexture& backgroundBuffer,
+    base::ArrayView<WaterEffectArea> areas,
+    int surfaceAnimationStep);
+
+private:
+  renderer::Renderer* mpRenderer;
+  renderer::Shader mShader;
+  renderer::CustomQuadBatch mBatch;
+  renderer::Texture mWaterSurfaceAnimTexture;
+  renderer::Texture mWaterEffectColorMapTexture;
+};
+
+} // namespace rigel::engine
